Update quick-start.md
As a first time user of this manual I was looking for the 'Payload Formats' menu item for a while and was not able to find it, while other buttons such as 'data' were available. Then I found out that in the step before at the 'Example Code', it is another screen (device) then the application screen, and that the device is not having a payload menu item. It was unclear to me by the current documentation that I have to return on a higher level of the application screen. I hope I made it more clear and less confusing for new users by this update.