Update faq.md
The first page I was looking for to open the gateway is the FAQ. Later on I found it at the bottom of https://www.thethingsnetwork.org/docs/gateways/gateway/ , but still it was unclear for me how to open it. Then I found out the video of Johan where he is opening it, so I added that part too.