Corrected some errors
Hi, I corrected some errors in the documentation and classified some things up. The casts are needed because if you shift a 8 bit variable by 24 for example the result will be 0. You can not assume the platform will automatically cast to 32 bit numbers for you. It is strongly recommended to use standard ints (uint8, uint16,...) but i do understand why this is not done considering the "Arduino" audience.