Bug Fix, incorrect LED state after disconnect with LED = On
I found a bug that occurs when you disconnect the things uno board from USB power supply when the LED is in on state.
When you reconnect the board the LED state starts in off state but the LoRa payloads return LED = On.  I think this is due to the pin not being correctly initialised as an output and so the digital read is reading a floating voltage on an input pin.
Adding the pinMode command in setup seemed to resolve this issue for me.