Update config.md
Using the steps from the Kerlink Wiki results in a more stable configuration than using the steps from the TTN docs. Alternatively the dota that is provided here should be updated to use the Kerlink SPF one from the Kerlink wiki, but with the TTN network server addresses configured already.