Kerlink firmware and packet forwarder doc update (#312)
* Update config.md

Using the steps from the Kerlink Wiki results in a more stable configuration than using the steps from the TTN docs. Alternatively the dota that is provided here should be updated to use the Kerlink SPF one from the Kerlink wiki, but with the TTN network server addresses configured already.

* Update config.md

* Update firmware-updates.md

A clearer description of how a Kerlink firmware update and factory reset should be performed. Also update to the latest firmware version.

* Update config.md

Almost a rewrite of the configuration page. We are now using the SPF packet forwarder as this one showed to be more stable than the TTN one. SPF is supported and actively updated by Kerlink, while the TTN one is outdated.

* Update cellular.md

* Update cellular.md

* Add files via upload

* Delete Kerlink ethernet.jpg

Wrong filename

* Add files via upload

* Add files via upload

* Update setup.md